seat: tweak focus to support multi-seat
This changes the case with no explicit target window to maintain the
current focus if the focused window is still visible rather than
focusing the topmost window.

With only one seat, this should have no effect. With more than one,
this allows for more independence in focus between the seats.